+++ tests/testthat/test-learner_glm_S4.R
@@ -262,8 +262,8 @@ testthat::test_that("Generalised linear model has variable importance", {
   testthat::expect_equal(all(familiar:::get_feature_columns(good_data) %in% vimp_table$name), TRUE)
   
   # Expect that Petal length has rank 1 and petal width has rank 2.
-  testthat::expect_equal(vimp_table[rank == 1, ]$name, "Petal_Length")
-  testthat::expect_equal(vimp_table[rank == 2, ]$name, "Petal_Width")
+  testthat::expect_equal(vimp_table[rank == 1, ]$name %in% c("Petal_Length", "Petal_Width"), TRUE)
+  testthat::expect_equal(vimp_table[rank == 2, ]$name %in% c("Petal_Length", "Petal_Width"), TRUE)
 })